+++ hostinfo/host/bare_views.py
@@ -19,10 +19,10 @@
 
 from django.shortcuts import render
 
-from .models import HostinfoException
+from .models import HostinfoException, getHostList
 
 from .views import criteriaFromWeb
-from .views import hostData, calcKeylistVals
+from .views import calcKeylistVals, hostData
 
 
 ################################################################################
@@ -35,6 +35,17 @@ def displayHost(request, hostname):
     """ Display a single host """
     d = hostData(request.user, [hostname], linker=getConfLinks)
     return render(request, 'bare/host.html', d)
+
+
+################################################################################
+def doHostCount(request, criturl):
+    """ Display the count of matching hosts """
+    criteria = criteriaFromWeb(criturl)
+    data = hostData(request, criteria)
+    try:
+        return render(request, 'bare/hostcount.html', data)
+    except HostinfoException as err:    # pragma: no cover
+        return render(request, 'bare/hostcount.html', {'error': err})
 
 
 ################################################################################
@@ -52,9 +63,18 @@ def doHostList(request, criturl):
 
 
 ################################################################################
-def doKeylist(request, key):
-    d = calcKeylistVals(key)
-    return render(request, 'bare/keylist.html', d)
+def doKeylist(request, key, criturl=None):
+    data = {}
+    if criturl:
+        criteria = criteriaFromWeb(criturl)
+        hostids = getHostList(criteria)
+        data['title'] = "Valuereport for %s: %s" % (key, " AND ".join(criteria))
+    else:
+        hostids = []
+        data['title'] = "Valuereport for %s" % key
+    data.update(calcKeylistVals(key, hostids))
+
+    return render(request, 'bare/keylist.html', data)
 
 
 ################################################################################